+++ api/views/session_test.py
@@ -64,100 +64,6 @@ class TestLoginView(TestCase):
             timeout=5,
         )
 
-    def test_login__invalid_code(self):
-        """Provided code must be valid."""
-        code = "7f06468085765cdc1578"
-
-        response = requests.Response()
-        response.status_code = status.HTTP_404_NOT_FOUND
-
-        with patch.object(
-            requests, "post", return_value=response
-        ) as requests_post:
-            res = self.client.post(
-                reverse("session-login"),
-                data={"code": code},
-                format="json",
-            )
-
-            assert res.status_code == status.HTTP_400_BAD_REQUEST
-            self._assert_request_github_access_token(requests_post, code)
-
-    def test_login__code_expired(self):
-        """Provided code must not expired."""
-        code = "7f06468085765cdc1578"
-
-        response = requests.Response()
-        response.status_code = status.HTTP_200_OK
-        response.encoding = "utf-8"
-        # pylint: disable-next=protected-access
-        response._content = json.dumps({"error": ""}).encode("utf-8")
-
-        with patch.object(
-            requests, "post", return_value=response
-        ) as requests_post:
-            res = self.client.post(
-                reverse("session-login"),
-                data={"code": code},
-                format="json",
-            )
-
-            assert res.status_code == status.HTTP_400_BAD_REQUEST
-            self._assert_request_github_access_token(requests_post, code)
-
-    def test_login__token_failure(self):
-        """Github did not access token did not get accepted by github."""
-        code = "7f06468085765cdc1578"
-
-        response_get = requests.Response()
-        response_get.status_code = status.HTTP_401_UNAUTHORIZED
-
-        with patch.object(
-            requests, "post", return_value=self.gh_access_token_response
-        ) as requests_post:
-            with patch.object(
-                requests, "get", return_value=response_get
-            ) as requests_get:
-                response = self.client.post(
-                    reverse("session-login"),
-                    data={"code": code},
-                    format="json",
-                )
-
-                assert response.status_code == status.HTTP_400_BAD_REQUEST
-
-                self._assert_request_github_access_token(requests_post, code)
-                self._assert_request_github_user(requests_get, "Bearer 123254")
-
-    def test_login__invalid_contributor_data(self):
-        """
-        Github did not provide data needed to log the user in.
-        """
-        code = "7f06468085765cdc1578"
-
-        response_get = requests.Response()
-        response_get.status_code = status.HTTP_200_OK
-        response_get.encoding = "utf-8"
-        # pylint: disable-next=protected-access
-        response_get._content = json.dumps({}).encode("utf-8")
-
-        with patch.object(
-            requests, "post", return_value=self.gh_access_token_response
-        ) as requests_post:
-            with patch.object(
-                requests, "get", return_value=response_get
-            ) as requests_get:
-                response = self.client.post(
-                    reverse("session-login"),
-                    data={"code": code},
-                    format="json",
-                )
-
-                assert response.status_code == status.HTTP_400_BAD_REQUEST
-
-                self._assert_request_github_access_token(requests_post, code)
-                self._assert_request_github_user(requests_get, "Bearer 123254")
-
     def test_login__new_contributor(self):
         """
         Login a new user with their existing github account and
